requirements.txt
django-annoying
django-cookie-law
django-appconf
lxml
slimit
cssmin
django_compressor
Pillow